firereach/backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-dotenv==1.0.1
groq==0.11.0
httpx==0.27.0
resend==2.0.0
pydantic==2.7.4
sse-starlette==2.1.0
email-validator==2.1.1
